backend/requirements.txt
Django==4.1.2
requests==2.28.1
django-environ==0.9.0
djangorestframework==3.14.0
dj-rest-auth==2.2.5
django-allauth==0.51.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.10.1
docker~=6.0.0
beautifulsoup4==4.11.1
copydetect==0.4.3
openai==0.25.0
multimetric
pylint
pycodestyle
mypy
